% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_se}
\alias{error_se}
\title{Calculate Standard Error of a Proportion}
\usage{
error_se(pct, N)
}
\arguments{
\item{pct}{numeric: Estimated proportion}

\item{N}{numeric: Number of Responses}
}
\value{
Returns a standard error value.
}
\description{
Calculate Standard Error of a Proportion
}
\examples{
# standard error for N=30: 33\%/67\%
x <- c(rep("Checked", 10), rep("Unchecked", 20))
N <- length(x)
pct <- prop.table(table(x))
error_se(pct, N)
}
\references{
\url{http://en.wikipedia.org/wiki/Margin_of_error} \cr
\url{http://en.wikipedia.org/wiki/Standard_error}
}
\seealso{
Other functions for estimating sampling error: 
\code{\link{error_me}()},
\code{\link{error_se_mean}()}
}
\concept{functions for estimating sampling error}
