% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_se_mean}
\alias{error_se_mean}
\title{Calculate Standard Error of the Mean}
\usage{
error_se_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector: Input numeric vector to be summarized}

\item{na.rm}{logical: If TRUE, NA (missing) values will be ignored.}
}
\value{
Returns a standard error value.
}
\description{
Calculate Standard Error of the Mean
}
\examples{
# generate some measurements & estimate std. error
x <- rlnorm(50, 3)
summary(x)
error_se_mean(x)

# introduce some missing values
nas <- rbinom(50, 1, 0.1)
x <- ifelse(nas, NA_integer_, x)
summary(x)
error_se_mean(x) # doesn't work
error_se_mean(x, na.rm = TRUE)
}
\references{
\url{http://en.wikipedia.org/wiki/Standard_error}
}
\seealso{
Other functions for estimating sampling error: 
\code{\link{error_me}()},
\code{\link{error_se}()}
}
\concept{functions for estimating sampling error}
