% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_mean_compare}
\alias{outlier_mean_compare}
\title{Compare averages with (newvar) vs. withoout (oldvar) outliers}
\usage{
outlier_mean_compare(df, oldvar, newvar, ...)
}
\arguments{
\item{df}{data frame with variables to compare}

\item{oldvar}{original variable without outliers removed}

\item{newvar}{variable with outliers removed}

\item{...}{optional grouping variables (use unquoted names)}
}
\description{
Compare averages with (newvar) vs. withoout (oldvar) outliers
}
\examples{
# see ?outlier_tukey
}
\seealso{
Other functions for identifying outliers: 
\code{\link{outlier_pct}()},
\code{\link{outlier_plot}()},
\code{\link{outlier_tukey}()}
}
\concept{functions for identifying outliers}
