% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_pct}
\alias{outlier_pct}
\title{Identify the percentage of values flagged as outliers}
\usage{
outlier_pct(df, ...)
}
\arguments{
\item{df}{data frame with outliers identified with df$is_outlier}

\item{...}{optional grouping variables (use unquoted names)}
}
\description{
Identify the percentage of values flagged as outliers
}
\examples{
# see ?outlier_tukey
}
\seealso{
Other functions for identifying outliers: 
\code{\link{outlier_tukey}()}
}
\concept{functions for identifying outliers}
