% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{Survey data for defining a population}
\format{A data frame with 877 rows and 4 columns (1 row per respondent)}
\usage{
pop
}
\description{
Survey data for defining a population
}
\seealso{
Other data: 
\code{\link{svy}}
}
\concept{data}
\keyword{datasets}
