% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{rake_weight}
\alias{rake_weight}
\title{Estimate rake weights for input survey data & population distributions}
\usage{
rake_weight(
  svy,
  pop,
  idvar,
  weightvar = "weight",
  print_name = "",
  cap = 20,
  force1 = TRUE,
  ...
)
}
\arguments{
\item{svy}{a survey data frame}

\item{pop}{a list that holds population distributions for the weighting
variables.}

\item{idvar}{name of varible that holds unique id}

\item{weightvar}{name of output weight variable}

\item{print_name}{header to print in summary (useful for log output)}

\item{cap}{cap argument of anesrake()}

\item{force1}{force1 argument of anesrake()}

\item{...}{other arguments passed to anesrake()}
}
\description{
This is a thin wrapper for \code{\link[anesrake]{anesrake}}. It prints a
summary and returns the survey dataset with a rake_wt variable appended.
}
\examples{
library(weights)
data(svy, pop)
svy <- svy$person

# determine target weight values
# the population is determined by a survey dataset which is itself weighted
wtvars <- c("age_weight", "income_weight", "race_weight")
pop_target <- sapply(wtvars, function(x) wpct(pop[[x]], pop$stwt))

# compare distributions for survey and population
pop_target
sapply(wtvars, function(x) wpct(svy[[x]]))

# run weighting
svy <- rake_weight(svy, pop_target, "Vrid")
sapply(wtvars, function(x) wpct(svy[[x]], svy$weight))
}
