% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svy}
\alias{svy}
\title{Survey sample data}
\format{A list containing 2 data frames
\describe{
  \item{person}{person-level demographic variables}
  \item{act}{person-activity-level activity data}
}}
\usage{
svy
}
\description{
Survey sample data
}
\concept{data}
\keyword{datasets}
