% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_me}
\alias{error_me}
\title{Calculate Margin of Error}
\usage{
error_me(std_error, confidence = 0.95)
}
\arguments{
\item{std_error}{numeric: Estimated standard error}

\item{confidence}{numeric: Confidence level (0 < confidence < 1)}
}
\value{
Returns a margin of error value.
}
\description{
Calculates margin of error for a given standard error at a given
confidence interval.
}
\examples{
# maximum margin of error (proportion) for N=100: 9.8\%
std_error <- error_se(0.5, 100)
error_me(std_error)
}
\references{
\url{http://en.wikipedia.org/wiki/Margin_of_error} \cr
\url{http://en.wikipedia.org/wiki/Standard_error}
}
\seealso{
Other functions for estimating sampling error: 
\code{\link{error_se_mean}()},
\code{\link{error_se}()}
}
\concept{functions for estimating sampling error}
