% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_plot}
\alias{outlier_plot}
\title{Make a plot of distributions with outliers identified}
\usage{
outlier_plot(df, var, grp, apply_log = FALSE)
}
\arguments{
\item{df}{data frame with outliers identified with df$is_outlier}

\item{var}{Unquoted name of variable to check}

\item{grp}{Unquoted name of variable to group by}
}
\description{
Make a plot of distributions with outliers identified
}
\examples{
# see ?outlier_tukey
}
\seealso{
Other functions for identifying outliers: 
\code{\link{outlier_pct}()},
\code{\link{outlier_tukey}()}
}
\concept{functions for identifying outliers}
